feat(art): add buy art endpoint and move hooks to features
refactor(frontend): relocate art and search hooks to feature directories
style(admin): improve dashboard revenue display with currency conversion
