+++ frontend/src/features/user/components/chat/chatArea/AddMemberModal.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import { useUnifiedSearch } from "../../../../../api/user/search/queries";
+import { useUnifiedSearch } from "../../../hooks/search/useUnifiedSearch";
 import { useSelector } from "react-redux";
 import { selectCurrentUser } from "../../../../../redux/selectors/userSelectors";
