+++ frontend/src/features/user/hooks/art/useDeleteCommentMutation.ts
@@ -1,0 +1,19 @@
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import apiClient from "../../../../api/axios";
+import { toast } from "react-hot-toast";
+
+export const useDeleteCommentMutation = (postId: string) => {
+  const queryClient = useQueryClient();
+  return useMutation({
+    mutationFn: (commentId: string) =>
+      apiClient.delete(`/api/v1/art/comments/${commentId}`),
+    onSuccess: () => {
+      toast.success("Comment deleted successfully");
+      queryClient.invalidateQueries({ queryKey: ["comments", postId] });
+    },
+    onError: (error) => {
+      console.error("Delete comment failed:", error);
+      toast.error("Failed to delete comment");
+    },
+  });
+};
