+++ frontend/src/features/user/pages/ArtPage.tsx
@@ -8,7 +8,8 @@ import { useUnlikePost } from "../hooks/art/useUnlikePost";
 import { useFavoritePost } from "../hooks/art/useFavoritePost";
 import { useUnfavoritePost } from "../hooks/art/useUnfavoritePost";
 import { useRelatedArtworks } from "../hooks/art/useRelatedArtworks";
-import { useBuyArtMutation, useDownloadArtMutation } from "../../../api/user/art/mutations";
+import { useBuyArtMutation } from "../hooks/art/useBuyArtMutation";
+import { useDownloadArtMutation } from "../hooks/art/useDownloadArtMutation";
 import ArtPageSkeleton from "../components/skeletons/ArtPageSkeleton";
 import ArtImageSection from "../components/art/details/ArtImageSection";
 import ArtActions from "../components/art/details/ArtActions";
